requirements.txt
# Core dependencies
requests>=2.25.0
Pillow>=8.0.0

# Optional dependencies for development
# pytest>=6.0
# pytest-cov>=2.0
# black>=21.0
# flake8>=3.8
# mypy>=0.800

# Optional dependencies for async support
# aiohttp>=3.7.0
